Move email-related code into the email file
