+++ plugins/wporg-5ftf/includes/pledge-meta.php
@@ -6,7 +6,7 @@
 namespace WordPressDotOrg\FiveForTheFuture\PledgeMeta;
 
 use WordPressDotOrg\FiveForTheFuture;
-use WordPressDotOrg\FiveForTheFuture\{ Pledge, PledgeForm, Contributor, XProfile };
+use WordPressDotOrg\FiveForTheFuture\{ Contributor, Email, Pledge, PledgeForm, XProfile };
 use WP_Post, WP_Error;
 
 defined( 'WPINC' ) || die();
@@ -274,14 +274,14 @@ function save_pledge( $pledge_id, $pledge ) {
 	save_pledge_meta( $pledge_id, $submitted_meta );
 
 	if ( filter_input( INPUT_POST, 'resend-pledge-confirmation' ) ) {
-		Pledge\send_pledge_confirmation_email(
+		Email\send_pledge_confirmation_email(
 			filter_input( INPUT_GET, 'resend-pledge-id', FILTER_VALIDATE_INT ),
 			get_page_by_path( 'for-organizations' )->ID
 		);
 	}
 
 	if ( filter_input( INPUT_POST, 'resend-contributor-confirmation' ) ) {
-		PledgeForm\send_contributor_confirmation_emails(
+		Email\send_contributor_confirmation_emails(
 			$pledge_id,
 			filter_input( INPUT_GET, 'resend-contributor-id', FILTER_VALIDATE_INT )
 		);
